Separate used modules and schema functions into their own files.
